% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_pull.R
\name{census_pull}
\alias{census_pull}
\title{Pulling Census ACS Data for DC}
\usage{
census_pull(geography, vars, year = 2018)
}
\arguments{
\item{geography}{Census geography, e.g. tract, blockgroup, state.}

\item{vars}{ACS variable names}

\item{year}{Year of 5-Year ACS, default = 2018}
}
\value{
sf projected in Maryland State Plane - espg 2248
}
\description{
Pulling Census ACS Data for DC
}
\examples{
census_pull("blockgroup", c("B01003_001E", "B19013_001E"))

}
